Added project
The project is empty. I did also download the shader graph editor which is a free Unity plug-in everyone has access to so it should be good. I did that while we were still narowing down what Unity version and I had gone through the repository steps up in till that project was needed. I do not know if we will us it in the end but it is there if we needed it.